4450 - repository/Country: unify returned columns for getOne, getMany
