+++ src/server/repository/assessmentCycle/country/getOne.ts
@@ -3,7 +3,9 @@ import { Objects } from 'utils/objects'
 import { Country, CountryIso } from 'meta/area'
 import { Assessment, Cycle } from 'meta/assessment'
 
-import { BaseProtocol, DB, Schemas } from 'server/db'
+import { BaseProtocol, DB } from 'server/db'
+
+import { getBaseQuery } from './_queries/getBaseQuery'
 
 export const getOne = async (
   props: { countryIso: CountryIso; assessment: Assessment; cycle: Cycle },
@@ -11,13 +13,5 @@ export const getOne = async (
 ): Promise<Country> => {
   const { countryIso, assessment, cycle } = props
 
-  const assessmentCycleName = Schemas.getNameCycle(assessment, cycle)
-
-  return client.oneOrNone<Country>(
-    `
-          select * from ${assessmentCycleName}.country c where c.country_iso = $1
-    `,
-    [countryIso],
-    Objects.camelize
-  )
+  return client.oneOrNone<Country>(getBaseQuery({ assessment, cycle, countryIso }), [countryIso], Objects.camelize)
 }
